Crate substrate_consensus_babe_primitives

Primitives for BABE.

Structs

BabeConfiguration
Configuration data used by the BABE consensus engine.

Epoch
BABE epoch information

NextEpochDescriptor
Information about the next epoch. This is broadcast in the first block of the epoch.

Enums

BabePreDigest
A BABE pre-runtime digest. This contains all data required to validate a block and for the BABE runtime module. Slots can be assigned to a primary (VRF based) and to a secondary (slot number based).

ConsensusLog
An consensus log item for BABE.

RawBabePreDigest
A raw version of BabePreDigest, usable on no_std.

Constants

BABE_ENGINE_ID
The ConsensusEngineId of BABE.

BABE_VRF_PREFIX
The prefix used by BABE for its VRF keys.

MEDIAN_ALGORITHM_CARDINALITY
How many blocks to wait before running the median algorithm for relative time This will not vary from chain to chain as it is not dependent on slot duration or epoch length.

PUBLIC_KEY_LENGTH
The length of the public key

VRF_OUTPUT_LENGTH
The length of the VRF output

VRF_PROOF_LENGTH
The length of the VRF proof

Traits

BabeApi
API necessary for block authorship with BABE.

CompatibleDigestItem
A digest item which is usable with BABE consensus.

Type Definitions

AuthorityId
A Babe authority identifier. Necessarily equivalent to the schnorrkel public key used in the main Babe module. If that ever changes, then this must, too.

AuthorityIndex
The index of an authority.

AuthorityPair
A Babe authority keypair. Necessarily equivalent to the schnorrkel public key used in the main Babe module. If that ever changes, then this must, too.

AuthoritySignature
A Babe authority signature.

BabeAuthorityWeight
The weight of an authority.

BabeBlockWeight
The weight of a BABE block.

SlotNumber
A slot number.